This commit to support exposing vendor specific options to application:
1) Define a structure defining the vendor specific options
2) Pass an array of vendor specific options needed with the call to OCDoResource
3) Vendor specific header options are checked against the range of acceptable vendor specific header options
4) Vendor specific header options are added to the pdu being sent
5) Vendor specific header options are parsed and added to OCEntityHandlerRequest and OCClientResponse structs
6) Rebased on to master
7) Adding a test case to occlient for sending vendor specific option
8) Adding The number of options to be sent and the ones received, also adding this infor to OCEntityHandlerRequest and OCClientResponse structs
9) Rebased on to master

+++ csdk/libcoap-4.1.1/bits.h
@@ -36,7 +36,7 @@ extern "C" {
  * @return @c -1 if @p bit does not fit into @p vec, @c 1 otherwise.
  */
 inline static int
-bits_setb(uint8_t *vec, size_t size, uint8_t bit) {
+bits_setb(uint8_t *vec, size_t size, uint16_t bit) {
   if (size <= (size_t)(bit >> 3))
     return -1;
 
@@ -56,7 +56,7 @@ bits_setb(uint8_t *vec, size_t size, uint8_t bit) {
  * @return @c -1 if @p bit does not fit into @p vec, @c 1 otherwise.
  */
 inline static int
-bits_clrb(uint8_t *vec, size_t size, uint8_t bit) {
+bits_clrb(uint8_t *vec, size_t size, uint16_t bit) {
   if (size <= (size_t)(bit >> 3))
     return -1;
 
@@ -75,7 +75,7 @@ bits_clrb(uint8_t *vec, size_t size, uint8_t bit) {
  * @return @c 1 if the bit is set, @c 0 otherwise.
  */
 inline static int
-bits_getb(const uint8_t *vec, size_t size, uint8_t bit) {
+bits_getb(const uint8_t *vec, size_t size, uint16_t bit) {
   if (size <= (size_t)(bit >> 3))
     return -1;
